Upgrade django requirement to 1.4.x
Some dependencies were not compatible with django 1.4.x
and had to be upgraded too.
Settings format changed a bit and was adapted. The most notable
change is the fact that MEDIA_URL and STATIC_URL have to
be distinct (they were the same so far).
The contrib.syndication API changed a bit and was adapted
accordingly.